requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-common==1.1.28
azure-core==1.35.1
azure-identity==1.25.0
azure-search-documents==11.6.0b1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.1
distro==1.9.0
fastapi==0.116.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.76
langchain-openai==0.3.33
langsmith==0.4.28
msal==1.33.0
msal-extensions==1.3.1
openai==1.108.0
orjson==3.11.3
packaging==25.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
zstandard==0.25.0
gunicorn==23.0.0
numpy==1.26.4